FS Cache: Split out GC file-scan
